UI details and date formatter
