+++ app/app/src/main/res/layout/activity_main.xml
@@ -9,25 +9,12 @@
     android:gravity="center"
     tools:context=".MainActivity">
 
-    <LinearLayout
+    <com.perficient.timeframechart.TimeFrameChatView
+        android:id="@+id/timeFrameChart"
         android:layout_gravity="center"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:orientation="vertical">
+        android:orientation="vertical"/>
 
-        <androidx.recyclerview.widget.RecyclerView
-            android:layout_gravity="center"
-            android:orientation="horizontal"
-            android:id="@+id/timeFrameRv"
-            app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"/>
-
-        <com.github.mikephil.charting.charts.BarChart
-            android:id="@+id/chart"
-            android:layout_width="match_parent"
-            android:layout_height="200dp" />
-
-    </LinearLayout>
 
 </LinearLayout>
