+++ app/settings.gradle
@@ -9,3 +9,4 @@ dependencyResolutionManagement {
 }
 rootProject.name = "Chart POC"
 include ':app'
+include ':timeFrameChart'
