+++ app/timeFrameChart/src/main/res/values/colors.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <color name="gray_char">#CFD2D5</color>
+    <color name="gray_calibrating">#4B4949</color>
+    <color name="transparent">#00FFFFFF</color>
+</resources>
